tracker/requirements.txt
psycopg2-binary
python-dotenv
structlog
requests